@@ -13,25 +13,25 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@vitejs/plugin-react": "^1.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "vite": "^2.8.4"
+    "@vitejs/plugin-react": "^1.3.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "vite": "^2.9.1"
   },
   "devDependencies": {
-    "@swc/core": "^1.2.144",
-    "@swc/jest": "^0.2.17",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
+    "@swc/core": "^1.2.163",
+    "@swc/jest": "^0.2.20",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.0.0",
     "@types/jest": "^27.4.1",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.14",
     "config": "*",
-    "eslint": "^8.9.0",
+    "eslint": "^8.12.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^27.5.1",
     "jest-watch-typeahead": "^1.0.0",
     "tsconfig": "*",
-    "typescript": "^4.5.5"
+    "typescript": "^4.6.3"
   }
 }
